Remodeled the charts
Added a single component for charts with the addition of type (for the charts we use currently).
TODO: Add options at a later date, not a priority
