+++ libs/upd/components/src/lib/charts/charts.component.css
@@ -1,5 +1,5 @@
-div.vh-50 {
-  height: 50vh;
+div.vh-45 {
+  height: 45vh;
 }
 
 div.mr-80 {
