--- libs/upd/components/src/lib/charts/gauge-chart/gauge-chart.component.html
@@ -1,29 +1,0 @@
-
-    <div #ContainerRef class="vh-45">
-
-<ngx-charts-gauge
-        [view]="[ContainerRef.offsetWidth, ContainerRef.offsetHeight]"
-        [legend]="showLegend"
-        [legendTitle]="legendTitle"
-        [legendPosition]="legendPosition"
-        [results]="single"
-        [animations]="animations"
-        [textValue]="gaugeTextValue"
-        [scheme]="colorScheme"
-        [min]="gaugeMin"
-        [max]="gaugeMax"
-        [units]="gaugeUnits"
-        [angleSpan]="gaugeAngleSpan"
-        [startAngle]="gaugeStartAngle"
-        [showAxis]="gaugeShowAxis"
-        [bigSegments]="gaugeLargeSegments"
-        [smallSegments]="gaugeSmallSegments"
-        [margin]="margin ? [marginTop, marginRight, marginBottom, marginLeft] : []"
-        [tooltipDisabled]="tooltipDisabled"
-        [showText]="showText"
-      [axisTickFormatting]="axisTickFormat"
-      [valueFormatting]="valueFormat"
-        (select)="onSelect($event)"
-      >
-      </ngx-charts-gauge>
-      </div>
