--- libs/upd/components/src/lib/charts/gauge-chart/gauge-chart.component.scss
@@ -1,4 +1,0 @@
-div.vh-45 {
-    height: 30vh;
-  }
-
