--- libs/upd/components/src/lib/charts/grouped-vertical-bar-chart/grouped-vertical-bar-chart.component.html
@@ -1,43 +1,0 @@
-<app-card [card-title]="title" [title-tooltip]="titleTooltip">
-  <div #ContainerRef class="vh-45">
-    <ngx-charts-bar-vertical-2d
-      [view]="[ContainerRef.offsetWidth, ContainerRef.offsetHeight]"
-      [scheme]="colorScheme"
-      [results]="barChart"
-      [animations]="animations"
-      [gradient]="gradient"
-      [xAxis]="showXAxis"
-      [yAxis]="showYAxis"
-      [legend]="showLegend"
-      [legendTitle]="legendTitle"
-      [legendPosition]="legendPosition"
-      [showXAxisLabel]="showXAxisLabel"
-      [showYAxisLabel]="showYAxisLabel"
-      [tooltipDisabled]="tooltipDisabled"
-      [xAxisLabel]="xAxisLabel"
-      [yAxisLabel]="yAxisLabel"
-      [showGridLines]="showGridLines"
-      [barPadding]="barPadding"
-      [groupPadding]="groupPadding"
-      [roundDomains]="roundDomains"
-      [roundEdges]="roundEdges"
-      [yScaleMax]="yScaleMax"
-      [noBarWhenZero]="noBarWhenZero"
-      [showDataLabel]="showDataLabel"
-      [trimXAxisTicks]="trimXAxisTicks"
-      [trimYAxisTicks]="trimYAxisTicks"
-      [rotateXAxisTicks]="rotateXAxisTicks"
-      [maxXAxisTickLength]="maxXAxisTickLength"
-      [maxYAxisTickLength]="maxYAxisTickLength"
-      [yAxisTickFormatting]="yAxisTickFormat"
-      (select)="onSelect($event)"
-    >
-      <ng-template #tooltipTemplate let-model="model">
-        <span class="tooltip-label"
-          >{{ model.series }} • {{ model.label }}</span
-        >
-        <span class="tooltip-val">{{ model.value.toLocaleString() }}</span>
-      </ng-template>
-    </ngx-charts-bar-vertical-2d>
-  </div>
-</app-card>
