--- libs/upd/components/src/lib/charts/grouped-vertical-bar-line-chart/grouped-vertical-bar-line-chart.component.html
@@ -1,57 +1,0 @@
-<app-card [card-title]="title" [title-tooltip]="titleTooltip" [h]="50">
-  <div #ContainerRef class="vh-50 mr-80">
-    <combo-grouped-vertical-bar-line-chart
-      [view]="[ContainerRef.offsetWidth, ContainerRef.offsetHeight]"
-      [scheme]="comboBarScheme"
-      [colorSchemeLine]="lineChartScheme"
-      [results]="barChart"
-      [animations]="animations"
-      [lineChart]="lineChart"
-      [tooltipDisabled]="tooltipDisabled"
-      [yAxisTickFormatting]="yLeftTickFormat"
-      [yLeftAxisScaleFactor]="yLeftAxisScale"
-      [yRightAxisScaleFactor]="yRightAxisScale"
-      [yRightAxisTickFormatting]="yRightTickFormat"
-      [gradient]="gradient"
-      [xAxis]="showXAxis"
-      [yAxis]="showYAxis"
-      [legend]="showLegend"
-      [legendTitle]="legendTitle"
-      [legendPosition]="legendPosition"
-      [showGridLines]="showGridLines"
-      [showXAxisLabel]="showXAxisLabel"
-      [showYAxisLabel]="showYAxisLabel"
-      [showRightYAxisLabel]="showRightYAxisLabel"
-      [xAxisLabel]="xAxisLabel"
-      [yAxisLabel]="yAxisLabel"
-      [yAxisLabelRight]="yAxisLabelRight"
-      [noBarWhenZero]="noBarWhenZero"
-      [barPadding]="barPadding"
-      [groupPadding]="groupPadding"
-      [colorLabelRight]="yAxisColourRight"
-      [colorLabelLeft]="yAxisColourLeft"
-      [roundEdges]="roundEdges"
-      [trimXAxisTicks]="trimXAxisTicks"
-      [trimYAxisTicks]="trimYAxisTicks"
-      [rotateXAxisTicks]="rotateXAxisTicks"
-      [maxXAxisTickLength]="maxXAxisTickLength"
-      [maxYAxisTickLength]="maxYAxisTickLength"
-      [showDataLabel]="showDataLabel"
-      [rangeFillOpacity]="rangeFillOpacity"
-      [autoScale]="autoScale"
-      [curve]="curve"
-      (select)="onSelect($event)"
-    >
-    </combo-grouped-vertical-bar-line-chart>
-  </div>
-
-  <div class="pt-3">
-    <app-accordion>
-      <app-data-table
-        [filter]="false"
-        [sort]="false"
-        [pagination]="false"
-      ></app-data-table>
-    </app-accordion>
-  </div>
-</app-card>
