--- libs/upd/components/src/lib/charts/horizontal-bar-chart/horizontal-bar-chart.component.html
@@ -1,36 +1,0 @@
-<app-card [card-title]="title" [title-tooltip]="titleTooltip" [h]="50">
-  <div #ContainerRef class="vh-45">
-    <ngx-charts-bar-horizontal
-      [view]="[ContainerRef.offsetWidth, ContainerRef.offsetHeight]"
-      [scheme]="colorScheme"
-      [results]="single"
-      [animations]="animations"
-      [gradient]="gradient"
-      [xAxis]="showXAxis"
-      [yAxis]="showYAxis"
-      [legend]="showLegend"
-      [legendTitle]="legendTitle"
-      [legendPosition]="legendPosition"
-      [showXAxisLabel]="showXAxisLabel"
-      [showYAxisLabel]="showYAxisLabel"
-      [tooltipDisabled]="tooltipDisabled"
-      [xAxisLabel]="xAxisLabel"
-      [yAxisLabel]="yAxisLabel"
-      [showGridLines]="showGridLines"
-      [barPadding]="barPadding"
-      [roundDomains]="roundDomains"
-      [roundEdges]="roundEdges"
-      [xScaleMax]="xScaleMax"
-      [noBarWhenZero]="noBarWhenZero"
-      [showDataLabel]="showDataLabel"
-      [trimXAxisTicks]="trimXAxisTicks"
-      [trimYAxisTicks]="trimYAxisTicks"
-      [rotateXAxisTicks]="rotateXAxisTicks"
-      [maxXAxisTickLength]="maxXAxisTickLength"
-      [maxYAxisTickLength]="maxYAxisTickLength"
-      [xAxisTickFormatting]="xAxisTickFormat"
-      (select)="onSelect($event)"
-    >
-    </ngx-charts-bar-horizontal>
-  </div>
-</app-card>
