+++ libs/upd/components/src/lib/charts/index.ts
@@ -1,11 +1,3 @@
-export * from './vertical-bar-chart/vertical-bar-chart.component';
 export * from './combo/combo-grouped-vertical-bar-line-chart/';
-export * from './gauge-chart/gauge-chart.component';
-export * from './grouped-vertical-bar-chart/grouped-vertical-bar-chart.component';
-export * from './vertical-bar-chart/vertical-bar-chart.component';
-export * from './grouped-vertical-bar-line-chart/grouped-vertical-bar-line-chart.component';
-export * from './horizontal-bar-chart/horizontal-bar-chart.component';
-export * from './line-chart/line-chart.component';
-export * from './stacked-vertical-bar-chart/stacked-vertical-bar-chart.component';
-export * from './pie-chart/pie-chart.component';
+export * from './charts.component';
 export * from './types';
