--- libs/upd/components/src/lib/charts/line-chart/line-chart.component.html
@@ -1,38 +1,0 @@
-<app-card [card-title]="title" [title-tooltip]="titleTooltip">
-  <div #ContainerRef class="vh-45">
-    <ngx-charts-line-chart
-      [view]="[ContainerRef.offsetWidth, ContainerRef.offsetHeight]"
-      [scheme]="colorScheme"
-      [results]="data"
-      [animations]="animations"
-      [legend]="showLegend"
-      [legendTitle]="legendTitle"
-      [legendPosition]="legendPosition"
-      [gradient]="gradient"
-      [xAxis]="showXAxis"
-      [yAxis]="showYAxis"
-      [showXAxisLabel]="showXAxisLabel"
-      [showYAxisLabel]="showYAxisLabel"
-      [xAxisLabel]="xAxisLabel"
-      [yAxisLabel]="yAxisLabel"
-      [autoScale]="autoScale"
-      [xScaleMin]="xScaleMin"
-      [xScaleMax]="xScaleMax"
-      [yScaleMin]="yScaleMin"
-      [yScaleMax]="yScaleMax"
-      [timeline]="timeline"
-      [showGridLines]="showGridLines"
-      [curve]="curve"
-      [rangeFillOpacity]="rangeFillOpacity"
-      [roundDomains]="roundDomains"
-      [tooltipDisabled]="tooltipDisabled"
-      [trimXAxisTicks]="trimXAxisTicks"
-      [trimYAxisTicks]="trimYAxisTicks"
-      [rotateXAxisTicks]="rotateXAxisTicks"
-      [maxXAxisTickLength]="maxXAxisTickLength"
-      [maxYAxisTickLength]="maxYAxisTickLength"
-      [yAxisTickFormatting]="yAxisTickFormat"
-      (select)="onSelect($event)"
-    ></ngx-charts-line-chart>
-  </div>
-</app-card>
