--- libs/upd/components/src/lib/charts/pie-chart/pie-chart.component.html
@@ -1,21 +1,0 @@
-<app-card [card-title]="title" [title-tooltip]="titleTooltip" [h]="50">
-  <div #ContainerRef class="vh-45">
-    <ngx-charts-pie-chart
-      [view]="[ContainerRef.offsetWidth, ContainerRef.offsetHeight]"
-      [scheme]="colorScheme"
-      [results]="data"
-      [animations]="animations"
-      [legend]="showLegend"
-      [legendTitle]="legendTitle"
-      [legendPosition]="legendPosition"
-      [explodeSlices]="explodeSlices"
-      [labels]="showLabels"
-      [doughnut]="doughnut"
-      [arcWidth]="arcWidth"
-      [gradient]="gradient"
-      [tooltipDisabled]="tooltipDisabled"
-      [tooltipText]="pieTooltipText"
-      (select)="onSelect($event)"
-    ></ngx-charts-pie-chart>
-  </div>
-</app-card>
