--- libs/upd/components/src/lib/charts/pie-chart/pie-chart.component.scss
@@ -1,3 +1,0 @@
-div.vh-45 {
-  height: 45vh;
-}
