--- libs/upd/components/src/lib/charts/stacked-vertical-bar-chart/stacked-vertical-bar-chart.component.html
@@ -1,34 +1,0 @@
-<app-card [card-title]="title" [title-tooltip]="titleTooltip">
-  <div #ContainerRef class="vh-45">
-    <ngx-charts-bar-vertical-stacked
-      [view]="[ContainerRef.offsetWidth, ContainerRef.offsetHeight]"
-      [scheme]="colorScheme"
-      [results]="data"
-      [animations]="animations"
-      [gradient]="gradient"
-      [tooltipDisabled]="tooltipDisabled"
-      [xAxis]="showXAxis"
-      [yAxis]="showYAxis"
-      [legend]="showLegend"
-      [legendTitle]="legendTitle"
-      [legendPosition]="legendPosition"
-      [showXAxisLabel]="showXAxisLabel"
-      [showYAxisLabel]="showYAxisLabel"
-      [xAxisLabel]="xAxisLabel"
-      [yAxisLabel]="yAxisLabel"
-      [showGridLines]="showGridLines"
-      [barPadding]="barPadding"
-      [roundDomains]="roundDomains"
-      [noBarWhenZero]="noBarWhenZero"
-      [yScaleMax]="yScaleMax"
-      [showDataLabel]="showDataLabel"
-      [trimXAxisTicks]="trimXAxisTicks"
-      [trimYAxisTicks]="trimYAxisTicks"
-      [maxXAxisTickLength]="maxXAxisTickLength"
-      [maxYAxisTickLength]="maxYAxisTickLength"
-      [yAxisTickFormatting]="yAxisTickFormat"
-      (select)="onSelect($event)"
-    >
-    </ngx-charts-bar-vertical-stacked>
-  </div>
-</app-card>
