--- libs/upd/components/src/lib/charts/vertical-bar-chart/vertical-bar-chart.component.html
@@ -1,40 +1,0 @@
-<app-card [card-title]="title" [title-tooltip]="titleTooltip" [h]="50">
-  <div #ContainerRef class="vh-45">
-    <ngx-charts-bar-vertical
-      [view]="[ContainerRef.offsetWidth, ContainerRef.offsetHeight]"
-      [scheme]="colorScheme"
-      [results]="single"
-      [animations]="animations"
-      [gradient]="gradient"
-      [xAxis]="showXAxis"
-      [yAxis]="showYAxis"
-      [legend]="showLegend"
-      [legendTitle]="legendTitle"
-      [legendPosition]="legendPosition"
-      [showXAxisLabel]="showXAxisLabel"
-      [showYAxisLabel]="showYAxisLabel"
-      [tooltipDisabled]="tooltipDisabled"
-      [xAxisLabel]="xAxisLabel"
-      [yAxisLabel]="yAxisLabel"
-      [showGridLines]="showGridLines"
-      [barPadding]="barPadding"
-      [roundDomains]="roundDomains"
-      [roundEdges]="roundEdges"
-      [yScaleMax]="yScaleMax"
-      [noBarWhenZero]="noBarWhenZero"
-      [showDataLabel]="showDataLabel"
-      [trimXAxisTicks]="trimXAxisTicks"
-      [trimYAxisTicks]="trimYAxisTicks"
-      [rotateXAxisTicks]="rotateXAxisTicks"
-      [maxXAxisTickLength]="maxXAxisTickLength"
-      [maxYAxisTickLength]="maxYAxisTickLength"
-      [yAxisTickFormatting]="yAxisTickFormat"
-      (select)="onSelect($event)"
-    >
-      <ng-template #tooltipTemplate let-model="model">
-        <span class="tooltip-label">{{ model.label }}</span>
-        <span class="tooltip-val">{{ model.value.toLocaleString() }}</span>
-      </ng-template>
-    </ngx-charts-bar-vertical>
-  </div>
-</app-card>
