--- libs/upd/components/src/lib/charts/vertical-bar-chart/vertical-bar-chart.component.spec.ts
@@ -1,24 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { VerticalBarChartComponent } from './vertical-bar-chart.component';
-
-describe('VerticalBarChartComponent', () => {
-  let component: VerticalBarChartComponent;
-  let fixture: ComponentFixture<VerticalBarChartComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [VerticalBarChartComponent],
-    }).compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(VerticalBarChartComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
