+++ libs/upd/components/src/lib/upd-components.module.ts
@@ -20,14 +20,7 @@ import { InputTextModule } from 'primeng/inputtext';
 import {
   ComboGroupedVerticalBarLineChartComponent,
   ComboSeriesVerticalComponent,
-  GaugeChartComponent,
-  GroupedVerticalBarLineChartComponent,
-  GroupedVerticalBarChartComponent,
-  HorizontalBarChartComponent,
-  LineChartComponent,
-  StackedVerticalBarChartComponent,
-  VerticalBarChartComponent,
-  PieChartComponent,
+  ChartsComponent,
 } from './charts';
 import { AlertComponent } from './alert/alert.component';
 import { AccordionComponent } from './accordion/accordion.component';
@@ -52,18 +45,11 @@ import { AccordionComponent } from './accordion/accordion.component';
     DateSelectorDropdownComponent,
     ComboGroupedVerticalBarLineChartComponent,
     ComboSeriesVerticalComponent,
-    GaugeChartComponent,
-    GroupedVerticalBarLineChartComponent,
-    GroupedVerticalBarChartComponent,
-    HorizontalBarChartComponent,
-    LineChartComponent,
-    StackedVerticalBarChartComponent,
-    VerticalBarChartComponent,
-    PieChartComponent,
     NavTabsComponent,
     DataCardComponent,
     AlertComponent,
     AccordionComponent,
+    ChartsComponent,
   ],
   exports: [
     NgxChartsModule,
@@ -76,18 +62,11 @@ import { AccordionComponent } from './accordion/accordion.component';
     DateSelectorDropdownComponent,
     ComboGroupedVerticalBarLineChartComponent,
     ComboSeriesVerticalComponent,
-    GaugeChartComponent,
-    GroupedVerticalBarLineChartComponent,
-    GroupedVerticalBarChartComponent,
-    HorizontalBarChartComponent,
-    LineChartComponent,
-    StackedVerticalBarChartComponent,
-    VerticalBarChartComponent,
-    PieChartComponent,
     NavTabsComponent,
     DataCardComponent,
     AlertComponent,
     AccordionComponent,
+    ChartsComponent,
   ],
 })
 export class UpdComponentsModule {}
