+++ libs/upd/views/overview/src/lib/overview-calldrivers/overview-calldrivers.component.html
@@ -1,6 +1,6 @@
 <div class="row mb-2 gx-2">
   <div class="col-lg-12 col-md-12 col-sm-12 h-100">
-    <app-grouped-vertical-bar-chart
+    <app-charts
       [title]="'Total calls by inquiry line'"
       [yAxisLabel]="'Number of calls (in thousands)'"
       [showXAxisLabel]="false"
@@ -8,8 +8,9 @@
       [displayLegend]="'below'"
       [groupPadding]="75"
       [showLegend]="true"
-      [barChart]="bar"
-    ></app-grouped-vertical-bar-chart>
+      [data]="bar"
+      [type]="'grouped-bar'"
+    ></app-charts>
   </div>
 </div>
