+++ libs/upd/views/pages/src/lib/pages-home/+state/pages-home.selectors.ts
@@ -18,5 +18,5 @@ export const getPagesHomeError = createSelector(
 
 export const getPagesHomeData = createSelector(
   getPagesHomeState,
-  (state: PagesHomeState) => state.data
+  (state: PagesHomeState) => [...state.data]
 );
